backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
scikit-learn==1.3.2
pandas==2.1.3
spotipy==2.23.0
redis==5.0.1
celery==5.3.6
httpx==0.25.2
python-multipart==0.0.6
supabase==2.3.0
sqlalchemy==2.0.23
asyncpg==0.29.0
